-> Revision of Version 1.1.0.2 - Added: Remove identifier from all FHIR resources - Added: EyeMaticsMdatPseudonymizer with Salt for unified identification - Added: CheckConsentMedicinalDataTask - Added: PseudonymizeDataTask - Added: Only EyeMatics associated profiles in Meta-Data are visible for every FHIR resource - Added: Replacement of Patients Meta-Data - Added: The DSF AllowList (Bundle) is used to create an Organization resource - Added: Affiliation of a Patient-Resource is visible under https://www.hl7.org/fhir/patient-definitions.html#Patient.managingOrganization as Reference via Organization resource - Added: MeasureReport Resource processing to ReadMedicinalDataTask and CreateDataBundleTask, associated with https://r4.fhir.space/measurereport-definitions.html#MeasureReport.reporter as Reference via Organization resource - Added: DiagnosticReport Resource processing to ReadMedicinalDataTask and CreateDataBundleTask, as belonging to EyeMatics CDS https://imi-ms.github.io/eyematics-kds/StructureDefinition-OphthalmicDiagnosticReport.html - Added: HashMaps for Observation, Patient, Medication regarding pseudonyms (globally / locally hashed) - Added: Test-Data-Generator (v1.2.0) with MeasureReportResource and DiagnosticReportResource - Added: New Errors: DATA_INSERT_FAILURE -> COMPLETE_INSERT_FAILURE and PARTIALLY_INSERT_FAILURE - Added: The duration time for the acknowledgement can be set dynamically via ORG_EYEMATICS_PROVIDE_WAIT_DURATION_PATIENTS_PER_BUNDLE variable - Changed: DSF Version from v1.7.1 to v1.9.0 - Changed: Logger Version from 2.0.16 to 2.0.9 (SLF4J) - Changed: Order of Resource Queries: Patient -> Consent -> fTTP -> MDAT -> ... - Changed: The number of FHIR Resource queries can be changed dynamically via Environment Variables - Changed: The number of BFs for an individual request can be changed dynamically via Environment Variables - Changed: Identifier verification is revised with an additional null check - Changed: Identifier replacement for all FHIR resources, only EyeMatics associated FHIR identifiers are held visible - Changed: KeyPair must be configured and provided. - Changed: KeyPair must be present, elsewise the Process Plugin will not start up - Changed: FHIR-Server URLs must be configured, elsewise the Process Plugin will not start up - Changed: Resource processing in CreateDataBundleTask - Changed: CheckBroadConsentTask -> ReadConsentDataTask - Changed: CreateDataBundleTask -> CreateDataBundlesTask - Changed: Blaze to 1.7.0 in docker-compose.yml - Changed: Shell-Script for insertion of test data - Changed: Patients distributed over locations can be found be same Identifier-Value (system: "https://ths-greifswald.de/gpas") - Changed: Consent Processing via Consent Processor - Changed: Patient and Resource Queries per Patient - Changed: Java-Stream-API is applied whenever possible (.parallel() when it is used to be thread-safe) - Changed: Patient- and Medical-Data Variables are overwritten in every processing step and deleted (null) when not used anymore - Fixed: Reduced the use of heap space by requesting and processing individual patient data and storing it in the DSFs database

+++ eyematics-dsf-process-plugin/src/main/java/org/eyematics/process/service/provide/DeleteProvideDataTask.java
@@ -9,8 +9,7 @@ import org.eyematics.process.constant.EyeMaticsGenericStatus;
 import org.eyematics.process.constant.ProvideConstants;
 import org.eyematics.process.utils.delegate.AbstractExtendedProcessServiceDelegate;
 import org.eyematics.process.utils.generator.DataSetStatusGenerator;
-import org.hl7.fhir.r4.model.Binary;
-import org.hl7.fhir.r4.model.IdType;
+import org.hl7.fhir.r4.model.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import dev.dsf.fhir.client.BasicFhirWebserviceClient;
@@ -27,16 +26,28 @@ public class DeleteProvideDataTask extends AbstractExtendedProcessServiceDelegat
 
     @Override
     protected void doExecute(DelegateExecution delegateExecution, Variables variables) throws BpmnError, Exception {
-        logger.info("-> Deleting the provided data");
+        logger.info("-> Deleting the provided data.");
         try {
-            IdType binaryId = new IdType(variables.getString(ProvideConstants.BPMN_PROVIDE_EXECUTION_VARIABLE_DATA_SET_REFERENCE));
-            if (!binaryId.isEmpty()) {
-                this.deletePermanently(binaryId);
-            }
+            Bundle referenceBundle =
+                    variables.getResource(ProvideConstants.BPMN_PROVIDE_EXECUTION_VARIABLE_DATA_SET_REFERENCE_BUNDLE);
+            if (referenceBundle == null) throw new Exception("Reference Bundle is null");
 
+            referenceBundle.getEntry()
+                    .stream()
+                    .filter(e -> e.getResource().getResourceType().equals(ResourceType.Bundle))
+                    .map(e -> (Bundle) e.getResource())
+                    .filter(b -> b.getEntry().size() == 1)
+                    .filter(b -> b.getEntry().get(0).getResource().getResourceType().equals(ResourceType.Basic))
+                    .map(b -> (Basic) b.getEntry().get(0).getResource())
+                    .forEach(b -> {
+                        IdType binaryId = new IdType(b.getSubject().getReference());
+                        if (!binaryId.isEmpty()) {
+                            this.deletePermanently(binaryId);
+                        }
+                    });
         } catch (Exception exception) {
             String errorMessage = exception.getMessage();
-            logger.error("Permanently deleting encrypted transferable data-set failed - {}", exception.getMessage());
+            logger.error("Permanently deleting encrypted transferable data-set failed - {}.", exception.getMessage());
             this.handleTaskError(EyeMaticsGenericStatus.DATA_DELETE_FAILURE, variables, errorMessage);
         }
     }
